- Saving objects as json rather than binary

+++ src/main/java/com/wealthy/machine/quote/DailyQuote.java
@@ -24,7 +24,7 @@ public interface DailyQuote extends Serializable, Comparable {
 
     Double getAvgPrice();
 
-    Double getVolume();
+    Integer getVolume();
 
     StockExchange getStockExchangeName();
